improve planForm
