+++ src/screens/planActivity/PlanForm.tsx
@@ -1,9 +1,14 @@
+import PlayButton, { TextInput } from 'components';
 import { Formik, FormikProps } from 'formik';
-import React from 'react';
-import { NavigationInjectedProps, withNavigation } from 'react-navigation';
-
+import { i18n } from 'locale';
 import { Plan } from 'models';
-import PlanFormControls from './PlanFormControls';
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+import { NavigationInjectedProps, withNavigation } from 'react-navigation';
+import { dimensions, palette } from 'styles';
+import { Icon, StyledText } from '../../components';
+import { PlanFormData } from './PlanForm';
+import { ShuffleButton } from './ShuffleButton';
 
 export interface PlanFormData {
   planInput: string;
@@ -13,22 +18,69 @@ const FORM_INITIAL_VALUES: PlanFormData = {
   planInput: '',
 };
 
-export class PlanForm extends React.PureComponent<NavigationInjectedProps> {
-  async createPlan(planInput: string) {
-    const { id } = this.props.navigation.getParam('student');
+interface Props {
+  onSubmit: (planFormData: PlanFormData) => Promise<void>;
+  plan?: Plan;
+}
 
-    const plan = await Plan.createPlan(id, planInput);
-
-    this.props.navigation.setParams({ plan });
-  }
-
-  onSubmit = ({ planInput }: PlanFormData) => this.createPlan(planInput);
-
-  renderFormControls = (props: FormikProps<PlanFormData>) => <PlanFormControls {...props} />;
+export class PlanForm extends React.PureComponent<Props> {
+  renderFormControls = ({ values, handleChange, submitForm }: FormikProps<PlanFormData>) => {
+    return (
+      <View style={styles.container}>
+        <View style={styles.inputContainer}>
+          <Icon name="emoticon" size={24} color={palette.textInputPlaceholder} />
+          {this.props.plan ? (
+            <StyledText style={styles.styledText}>{values.planInput}</StyledText>
+          ) : (
+            <TextInput
+              style={styles.textInput}
+              placeholder={i18n.t('planActivity:planNamePlaceholder')}
+              value={values.planInput}
+              onChangeText={handleChange('planInput')}
+              onEndEditing={submitForm}
+            />
+          )}
+        </View>
+        <View style={styles.buttonContainer}>
+          <ShuffleButton disabled={!this.props.plan} />
+          <PlayButton plan={this.props.plan} disabled={!this.props.plan} size={36} />
+        </View>
+      </View>
+    );
+  };
 
   render() {
-    return <Formik initialValues={FORM_INITIAL_VALUES} onSubmit={this.onSubmit} render={this.renderFormControls} />;
+    return (
+      <Formik initialValues={FORM_INITIAL_VALUES} onSubmit={this.props.onSubmit} render={this.renderFormControls} />
+    );
   }
 }
 
-export default withNavigation(PlanForm);
+const styles = StyleSheet.create({
+  container: {
+    height: 56,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    paddingLeft: dimensions.spacingLarge,
+    paddingRight: dimensions.spacingBig,
+    borderBottomColor: palette.backgroundAdditional,
+    borderBottomWidth: 1,
+  },
+  inputContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    maxWidth: '40%',
+  },
+  buttonContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  styledText: {
+    color: palette.textBlack,
+    marginLeft: dimensions.spacingSmall,
+  },
+  textInput: {
+    marginLeft: dimensions.spacingSmall,
+  },
+});
