+++ src/screens/planActivity/__tests__/PlanForm.test.tsx
@@ -2,11 +2,14 @@ import { shallow } from 'enzyme';
 import React from 'react';
 
 import { navigationMock } from 'helpers';
+import { Plan } from 'models';
 import { PlanForm } from '../PlanForm';
 
 describe('PlanFormControls', () => {
   const props = {
     navigation: navigationMock,
+    onSubmit: jest.fn(),
+    plan: new Plan(),
   };
 
   it('renders correctly', () => {
