--- src/screens/planActivity/__tests__/PlanFormControls.test.tsx
@@ -1,22 +1,0 @@
-import { shallow } from 'enzyme';
-import { navigationMock } from 'helpers';
-import React from 'react';
-import { PlanFormControls } from '../PlanFormControls';
-
-describe('PlanFormControls', () => {
-  const props = {
-    handleChange: jest.fn(),
-    values: {
-      planInput: 'Music',
-    },
-    submitForm: jest.fn(),
-    isSubmitting: false,
-    navigation: navigationMock,
-  } as any;
-
-  it('renders correctly', () => {
-    const planInput = shallow(<PlanFormControls {...props} />);
-
-    expect(planInput).toMatchSnapshot();
-  });
-});
